% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{sn_set_api_token}
\alias{sn_set_api_token}
\title{Configure an API bearer token}
\usage{
sn_set_api_token(token = NULL)
}
\arguments{
\item{token}{Token string, or \code{NULL} to clear the option.}
}
\value{
The previous configured token, invisibly.
}
\description{
Configure an API bearer token
}
